requirements.txt
# ── Core framework ─────────────────────────────────────────────────────────────
Django==5.0.6
djangorestframework==3.15.2
django-environ==0.11.2          # .env → settings bridge

# ── Async task queue ───────────────────────────────────────────────────────────
celery==5.4.0
redis==5.0.7                    # Celery broker + result backend

# ── Database ───────────────────────────────────────────────────────────────────
psycopg2-binary==2.9.9          # PostgreSQL driver (Docker); SQLite used locally

# ── XML handling (local, zero cost) ───────────────────────────────────────────
lxml==5.2.2

# ── Gemini / Google AI ─────────────────────────────────────────────────────────
google-genai                     # Modern unified SDK for Gemini/Imagen/Veo

# ── Dev / utilities ────────────────────────────────────────────────────────────
pydantic>=2.0                   # Blueprint schema validation (schemas.py)
requests==2.32.3
python-dotenv==1.0.1
moviepy==1.0.3                  # Video stitching engine
pillow>=10.0.0                  # Image fallback / handling
gTTS==2.5.1                     # Local TTS engine for audio placeholder
pytest>=8.0.0                   # Testing framework
pytest-django>=4.8.0            # Django plugin for pytest
django-stubs                    # Type stubs for Django
